backend/requirements.txt
fastapi
uvicorn
motor
python-jose
bcrypt
python-dotenv
pydantic
email-validator
requests
aiohttp
gunicorn
python-multipart